back-task-gestion/pyproject.toml
[project]
name = "back-task-gestion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bcrypt>=4.0.1,<5.0.0",
    "fastapi>=0.118.2",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.12.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.37.0",
]
